requirements.txt
Flask==3.0.0
pandas==2.1.4
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.24